requirements.txt
appdirs==1.4.3
asgiref==3.2.5
astroid==2.3.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.41.0
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-authtoken==1.2.4
djangorestframework==3.11.0
drf-extensions==0.6.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pathspec==0.7.0
psycopg2==2.7.7
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.11.2
